test(item-sliding): migrate scroll-target
